coding-templates/python/dwave-hello-service/requirements.txt
loguru
dwave-ocean-sdk
